requirements.txt
requests>=2.28.0,<3.0.0
Pillow>=9.0.0,<11.0.0
ebooklib>=0.18,<0.19
fake-useragent>=1.5.0,<2.0.0
tqdm>=4.65.0,<5.0.0
beautifulsoup4>=4.12.0,<5.0.0
urllib3>=1.26.0,<3.0.0
packaging>=23.1,<25.0
# HEIC图片格式支持
pillow-heif>=0.15.0